Create baseline.md
This is a first draft that I started editing for 9.5 Patch 5 work items 134198, 134048, 134201, 134199. It is not ready for review at all, being a copy of bigfix query rest api resource, it still contains parts that belonged to bigfix query. Just to see it show up and check that nothing gets broken.